540 concept sets user specific (#63)
* Add migrations for user id and shared in terminology

* update migrations

* update migration

* update concept set cruds for shared

+++ services/alp-terminology-svc/src/module/concept-set/concept-set.service.ts
@@ -9,6 +9,7 @@ import { Request } from 'express';
 import { ConceptService } from '../concept/concept.service';
 import { MeilisearchAPI } from '../../api/meilisearch-api';
 import { SystemPortalAPI } from 'src/api/portal-api';
+import { Brackets } from 'typeorm';
 
 @Injectable()
 export class ConceptSetService {
@@ -57,6 +58,8 @@ export class ConceptSetService {
     const conceptSets = await dataSource
       .getRepository(ConceptSet)
       .createQueryBuilder()
+      .where({ createdBy: this.userId })
+      .orWhere({ shared: true })
       .orderBy('name')
       .getMany();
 
@@ -91,8 +94,15 @@ export class ConceptSetService {
     const dataSource = await this.getDataSource();
     const conceptSet = await dataSource
       .getRepository(ConceptSet)
-      .createQueryBuilder()
-      .where({ id: conceptSetId })
+      .createQueryBuilder('conceptSet')
+      .where('conceptSet.id = :id', { id: conceptSetId })
+      .andWhere(
+        new Brackets((qb) => {
+          qb.where('conceptSet.createdBy = :createdBy', {
+            createdBy: this.userId,
+          }).orWhere('conceptSet.shared = :shared', { shared: true });
+        }),
+      )
       .getOne();
     const conceptIds = conceptSet.concepts.map((c) => c.id);
     const conceptService = new ConceptService(this.request);
@@ -127,6 +137,7 @@ export class ConceptSetService {
       .update(ConceptSet)
       .set(conceptSetData)
       .where({ id: conceptSetId })
+      .andWhere({ createdBy: this.userId })
       .execute();
     return conceptSetId;
   }
@@ -138,6 +149,7 @@ export class ConceptSetService {
       .delete()
       .from(ConceptSet)
       .where({ id: conceptSetId })
+      .andWhere({ createdBy: this.userId })
       .execute();
     return conceptSetId;
   }
